Refactor some stuff.
Remove some unnecessary uses of boolean and fixed-width integer types.

+++ src/exercise110.c
@@ -6,13 +6,12 @@
  * information, please refer to the accompanying "UNLICENCE" file.
  */
 
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
 int main(void)
 {
-  int16_t character = 0;
+  int character = 0;
   while ((character = getchar()) != EOF) {
     switch (character) {
       case '\t':
